+++ .github/workflows/prerelease.yml
@@ -11,7 +11,7 @@ jobs:
     - uses: actions/checkout@v2
     - uses: actions/setup-node@v3.0.0
       with:
-        node-version: '12.x'
+        node-version: '14'
         registry-url: 'https://registry.npmjs.org'
     - run: npm ci --no-progress --production
     - run: npm version --no-push --no-git-tag-version --yes ${{ github.event.release.tag_name }}
