+++ src/utils/getNameError.ts
@@ -4,7 +4,7 @@
  * @param   {object} results Results from validate-npm-package-name
  * @returns {string}         Error/warning message
  */
-const getNameError = (results) => {
+export const getNameError = (results: any): any => {
   // Errors are returned for names that were never valid
   if (results.errors && results.errors.length > 0) {
     return results.errors[0];
@@ -18,5 +18,3 @@ const getNameError = (results) => {
   // Ensure that an error message is returned in any case
   return 'name invalid';
 };
-
-module.exports = getNameError;
