+++ src/validators/format.ts
@@ -2,14 +2,6 @@ import semver from 'semver';
 import {PackageJson} from 'type-fest';
 
 /**
- * Determines whether or not the string is lowercase
- *
- * @param name Name
- * @return True if the string is lowercase or is missing. False if it is not.
- */
-export const isLowercase = (name: string): boolean => name === name.toLowerCase();
-
-/**
  * Determines whether or not the node's value is a valid semantic version
  *
  * @param packageJsonData Valid JSON
